fix: change routes and validator according to the check-list
