+++ middlewares/validator.js
@@ -1,0 +1,70 @@
+const { celebrate, Joi } = require('celebrate');
+const validator = require('validator');
+
+const emailValidation = (value) => {
+  const result = validator.isEmail(value);
+  if (result) {
+    return value;
+  }
+  throw new Error('Email validation failed');
+};
+
+const urlValidation = (value) => {
+  const result = validator.isURL(value);
+  if (result) {
+    return value;
+  }
+  throw new Error('URL validation failed');
+};
+
+const signupValidation = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30),
+    email: Joi.string().required().email(),
+    password: Joi.string().required(),
+  }),
+});
+
+const signinValidation = celebrate({
+  body: Joi.object().keys({
+    email: Joi.string().required().email(),
+    password: Joi.string().required(),
+  }),
+});
+
+const createMovieValidation = celebrate({
+  body: Joi.object().keys({
+    country: Joi.string().required(),
+    director: Joi.string().required(),
+    duration: Joi.number().required(),
+    year: Joi.string().required(),
+    description: Joi.string().required(),
+    image: Joi.string().required().custom(urlValidation),
+    trailer: Joi.string().required().custom(urlValidation),
+    nameRU: Joi.string().required(),
+    nameEN: Joi.string().required(),
+    thumbnail: Joi.string().required().custom(urlValidation),
+    movieId: Joi.string().required(),
+  }),
+});
+
+const deleteMovieValidation = celebrate({
+  params: Joi.object().keys({
+    movieId: Joi.string().length(24).hex(),
+  }),
+});
+
+const updateUserDataValidation = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().required().custom(emailValidation),
+  }),
+});
+
+module.exports = {
+  signupValidation,
+  signinValidation,
+  createMovieValidation,
+  deleteMovieValidation,
+  updateUserDataValidation,
+};
