+++ routes/movies.js
@@ -1,14 +1,5 @@
 const router = require('express').Router();
 const validator = require('validator');
-const { celebrate, Joi } = require('celebrate');
-
-const urlValidation = (value) => {
-  const result = validator.isURL(value);
-  if (result) {
-    return value;
-  }
-  throw new Error('Url не валидный');
-};
 
 const {
   getMovies,
@@ -16,24 +7,13 @@ const {
   deleteMovie,
 } = require('../controllers/movies');
 
+const {
+  createMovieValidation,
+  deleteMovieValidation,
+} = require('../middlewares/validator');
+
 router.get('/movies', getMovies);
-router.post('/movies', celebrate({
-  body: Joi.object().keys({
-    country: Joi.string().required(),
-    director: Joi.string().required(),
-    duration: Joi.number().required(),
-    year: Joi.string().required(),
-    description: Joi.string().required(),
-    image: Joi.string().required().custom(urlValidation),
-    trailer: Joi.string().required().custom(urlValidation),
-    nameRU: Joi.string().required(),
-    nameEN: Joi.string().required(),
-  }),
-}), createMovie);
-router.delete('/movies/:movieId', celebrate({
-  params: Joi.object().keys({
-    movieId: Joi.string().length(24).hex(),
-  }),
-}), deleteMovie);
+router.post('/movies', createMovieValidation, createMovie);
+router.delete('/movies/:movieId', deleteMovieValidation, deleteMovie);
 
 module.exports = router;
